// src/modelValidation/dish.schema.ts
import { DishStatusValues } from '@/constants/type'
import z from 'zod'

export const CreateDishBody = z.object({
  name: z.string().min(1,{error : 'Tối thiểu 1 kí tự'}).max(256,{error :'Tối đa 256 kí tự'}),
  price: z.number({error : 'Không được để trống'}).positive({error : 'Giá phải lớn hơn 0'}),
  description: z.string().max(10000),
  image: z.url().optional(),
  status: z.enum(DishStatusValues).optional()
})

export type CreateDishBodyType = z.infer<typeof CreateDishBody>

export const DishSchema = z.object({
  id: z.number(),
  name: z.string(),
  price: z.coerce.number(),
  description: z.string(),
  image: z.string(),
  status: z.enum(DishStatusValues),
  createdAt: z.date(),
  updatedAt: z.date()
})

export const DishRes = z.object({
  data: DishSchema,
  message: z.string()
})

export type DishResType = z.TypeOf<typeof DishRes>

export const DishListRes = z.object({
  data: z.array(DishSchema),
  message: z.string()
})

export type DishListResType = z.TypeOf<typeof DishListRes>

export const UpdateDishBody = CreateDishBody
export type UpdateDishBodyType = CreateDishBodyType
export const DishParams = z.object({
  id: z.coerce.number()
})
export type DishParamsType = z.TypeOf<typeof DishParams>